Update styles for homepage and login form
The SCSS stylesheets have been updated to improve the visual presentation of the homepage for logged out users, and restyle the login form. This includes central alignment of the site title on the homepage, adjusting padding and margins, and implementing new stylistic changes to the login form's submit button. The updates are aimed at creating a more engaging user experience and enhancing the overall look and feel of the site.