Update styling for header and body elements
Enhanced the styling configurations for header and body elements within 'c5connections-theme'. For the header, display rules and margin adjustments were added for 'c5-logo' and 'site-title' in relation to logged-in state. For the body, a default page template with a new background gradient and updates to the 'primary' id were implemented.